+++ packages/core/src/node/perm_group/mod.rs
@@ -4,7 +4,7 @@ use crate::node::Node;
 
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub struct PermGroupNode {
-    pub node_id: u128,
+    pub node_id: u64,
     pub node_iden: String,
     pub name: String,
     pub description: String,
@@ -14,7 +14,7 @@ pub struct PermGroupNode {
 }
 
 impl Node for PermGroupNode {
-    fn get_node_id(&self) -> u128 {
+    fn get_node_id(&self) -> u64 {
         self.node_id
     }
